feat : resume 컨트롤러 추가
Former-commit-id: d1475dc9a25a64c0d7c3b4d983a9f4771af7137b [formerly 9fb1873b9f6c2aaafeaf9ac07889383664c5b269]
Former-commit-id: c01a1c293b1b6fe33c4c59ff55353dd5118b266e

+++ back-end/src/main/java/com/ssafy/ilbok/Repository/RegionsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RegionsRepository extends JpaRepository<Regions, Long> {
+public interface RegionsRepository extends JpaRepository<Regions, Integer> {
     List<Regions> findAll();
 }
